Move builder and outputs to spin module
